+++ src/domain/communication/room/dto/room.dto.remove.message.reaction.ts
